Fix rpm/yum installed packages detecting, fix lots of small things. (#540)

+++ deployment/ceph-tools/install/common/install-itsdangerous.sh
@@ -1,12 +1,18 @@
 set -eu
 
-file="python-itsdangerous-0.23-2.el7.noarch.rpm"
-rm -f "$file"
+itsdangerous_installed=`rpm -qa | grep itsdangerous`
 
-echo "=> curl -Ov http://mirror.centos.org/centos/7/extras/x86_64/Packages/$file"
-curl -Ov "http://mirror.centos.org/centos/7/extras/x86_64/Packages/$file"
+if [ ! -z "$itsdangerous_installed" ]; then
+	echo "=> python-itsdangerous installed, skipped"
+else
+	file="python-itsdangerous-0.23-2.el7.noarch.rpm"
+	rm -f "$file"
 
-echo "=> sudo rpm -i $file"
-sudo rpm -i "$file"
+	echo "=> curl -Ov http://mirror.centos.org/centos/7/extras/x86_64/Packages/$file"
+	curl -Ov "http://mirror.centos.org/centos/7/extras/x86_64/Packages/$file"
 
-rm -f "$file"
+	echo "=> sudo rpm -i $file"
+	sudo rpm -i "$file"
+
+	rm -f "$file"
+fi
